requirements.txt
alabaster==0.7.12
attrs==21.2.0
Babel==2.9.1
certifi==2021.10.8
charset-normalizer==2.0.7
coverage==6.1.2
cycler==0.11.0
docutils==0.17.1
idna==3.3
imagesize==1.3.0
iniconfig==1.1.1
Jinja2==3.0.3
kiwisolver==1.3.2
MarkupSafe==2.0.1
matplotlib==3.4.3
multiprocess==0.70.12.2 
mypy==0.910
mypy-extensions==0.4.3
numpy==1.21.3
opt-einsum==3.3.0
packaging==21.0
pandas==1.3.4
Pillow==8.4.0
pip-licenses==3.5.3
pluggy==1.0.0
PTable==0.9.2
py==1.10.0
Pygments==2.10.0
pyparsing==3.0.0
pyro-api==0.1.2
pyro-ppl==1.7.0
pytest==6.2.5
pytest-cov==3.0.0
python-dateutil==2.8.2
pytz==2021.3
requests==2.26.0
scipy==1.7.3
six==1.16.0
snowballstemmer==2.1.0
Sphinx==4.3.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
toml==0.10.2
tomli==1.2.2
torch==1.10.0
tqdm==4.62.3
typing-extensions==3.10.0.2
urllib3==1.26.7
